+++ src/main/java/com/hmshop/application/model/dto/ProductDTO.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -16,8 +17,8 @@ public class ProductDTO {
     private String id;
     private String name;
     private  String description;
-    private long price;
-    private long salePrice;
+    private BigDecimal price;
+    private BigDecimal salePrice;
     private long totalSold;
     private int status;
     private List<String> images;
